FIX: Fix removing disordered phase candidates (#16)
Fixes a regression introduced in #11, where `filter_phases` was introduced to remove phases that cannot be stable or phases where the disordered part is given as a candidate phase. The change in #11 made it so the disordered phase was not a part of the list of `solid_phases` that could form during the solidification simulations. With the disordered phase name not included in the `solid_phases`, the `order_disorder_eq_phases` function that renamed the ordered phases with disordered configurations to their disordered counterpart would lead to the phase amount of disordered configurations to _not_ be saved.

This PR fixes that and adds a parameterized test case that tests that the disordered phase always is accounted for regardless of whether or not the ordered phase is in the candidate phases or not.