ENH: Generalize phase composition support (#35)
Until now we have only tracked liquid phase compositions in `x_liquid`. Now we are tracking phase compositions of all phases. Note that we don't yet support phase multiplicities, so miscibility gaps _will_ give potentially meaningless results. This change also introduces tracking of the phase composition of the dependent component as well, for convenience.

We are keeping `x_liquid` around for now for backwards compatibility, but eventually it should probably be removed and superseded by `phase_compositions` (since the liquid phase is in the `phase_compositions` dictionary).